Never call srand().
Modern versions of Perl, particularly since 5.8.x (which this module
targets), will automatically seed `srand()` with a reasonably
entrophic result (either taken from /dev/urandom, or a mix of
time, process ID, and memory allocation). These seeds are *more*
random than the seed which Pwgen uses, which is only the current
epoch time (which is already known) and a sequence number.

Consequently, calling srand() directly results in *less* random
passwords, and multiple calls to srand() actually make them less
random still (see `perldoc -f srand` for a more in-depth discussion
on this).

This patch removes the calls to srand() entirely.